requirements.txt
pandas
requests
python-dotenv
textstat
google-genai
html
re
numpy
scipy 
seaborn
